requirements.txt
advent-of-code-data~=2.0.1
click~=8.1.7
rich~=13.7.0
pyyaml
